Make a new requirements.txt file for pytest and a script to install every lambda we have made as a pip package
