+++ lambda/fetchers/get_docs/setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup
+
+setup(name='get_docs', version='1.0.0', packages=['.'], exclude=["lambda_function"])
