+++ lambda/ingest/setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup
+
+setup(name='ingest', version='1.0.0', packages=['.'], exclude=["lambda_function"])
