+++ lambda/requirements.txt
@@ -1,0 +1,20 @@
+# The lambdas don't need boto in their requirements file because it's already installed on amazon lambda.
+atomicwrites==1.3.0
+attrs==19.1.0
+boto3==1.9.178
+botocore==1.12.178
+docutils==0.14
+importlib-metadata==0.18
+jmespath==0.9.4
+more-itertools==7.1.0
+packaging==19.0
+pluggy==0.12.0
+py==1.8.0
+pyparsing==2.4.0
+pytest==4.6.3
+python-dateutil==2.8.0
+s3transfer==0.2.1
+six==1.12.0
+urllib3==1.25.3
+wcwidth==0.1.7
+zipp==0.5.1
