+++ lambda/setup.sh
@@ -1,0 +1,5 @@
+pip install -r requirements.txt
+pip install layers/python/
+pip install batch_job_aurora/
+pip install ingest/
+pip install fetchers/get_docs/
